fix: page.tsx 서버 컴포넌트로 변경 후 클라이언트 컴포넌트 불러오기
