+++ src/app/(auth)/login/page.tsx
@@ -5,7 +5,8 @@ import styled from "@emotion/styled";
 import Heading from "@/components/Text/Heading";
 import { HStack, Separator } from "@chakra-ui/react";
 import Image from "next/image";
-import { socialLogin, checkUser } from "@/app/(auth)/actions";
+import { checkUser } from "@/app/(auth)/actions";
+import { socialLogin } from "@/app/(auth)/clientActions";
 import { Modal } from "@/components/modal/Modal";
 import LoginSignup from "@/components/auth/LoginSignup";
 import Text from "@/components/Text/Text";
