+++ src/hooks/useUser.ts
@@ -1,11 +1,12 @@
 import { useQuery } from "@tanstack/react-query";
-import { supabase } from "@/utils/supabase/client";
+import { createClient } from "@/utils/supabase/client";
 
 export const useUser = (id: number | null) => {
   const { data, isLoading, error } = useQuery({
     queryKey: ["profile-user-data", id],
     queryFn: async () => {
       if (id === null) return null;
+      const supabase = createClient();
       const { data, error } = await supabase
         .from("profiles")
         .select("*")
