+++ src/utils/supabase/client.ts
@@ -7,5 +7,3 @@ export function createClient() {
     process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY!,
   );
 }
-
-export const supabase = createClient();
