--- src/utils/uploadFile.ts
@@ -1,8 +1,0 @@
-import { supabase } from "@/utils/supabase/client";
-
-export const uploadFile = async (file: File) => {
-  const { data, error } = await supabase.storage
-    .from("my-bucket")
-    .upload(file.name, file);
-  return { data, error };
-};
