+++ src/utils/user/user.ts
@@ -1,6 +1,7 @@
-import { supabase } from "@/utils/supabase/client";
+import { createClient } from "@/utils/supabase/client";
 
 export const checkUser = async (): Promise<boolean> => {
+  const supabase = createClient();
   const {
     data: { user },
     error,
